0000 | creates a new wallet section, randomizes the cash in your wallet, initial cash on hand is parsed from this config now
